Tests should check the datatypes of numbers
Comparisons between double and log should not be equal

+++ array/src/commonMain/kotlin/array/builtins/math_functions.kt
@@ -368,7 +368,7 @@ class MulAPLFunction : APLFunctionDescriptor {
                 pos,
                 a,
                 { x -> x.sign.toLong().makeAPLNumber() },
-                { x -> x.sign.makeAPLNumber() },
+                { x -> x.sign.toLong().makeAPLNumber() },
                 { x -> x.signum().makeAPLNumber() })
         }
